% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteforce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Solve the knapsack problem using a brute-force algorithm}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A dataframe called knapsack_objects with items with two columns: v- values and w-weights per items.}

\item{W}{a numeric variable which represents the size of the knapsack}

\item{parallel}{a variable with boolean values (TRUE or FALSE). If set to true, the function should parallelize.}
}
\value{
A list with two elements: the maximum value and a vector of selected items.
}
\description{
Solve the knapsack problem using a brute-force algorithm
}
