% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Solve the knapsack problem using a greedy algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe called knapsack_objects with items with two columns: v- values and w-weights per items.}

\item{W}{a numeric variable which represents the size of the knapsack}
}
\value{
A list with two elements: the maximum value and a vector of selected item indices
}
\description{
Solve the knapsack problem using a greedy algorithm
}
