% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicProgramming.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solve the knapsack problem using dynamic programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A dataframe called knapsack_objects with items with two columns: v- values and w-weights per items.}

\item{W}{a numeric variable which represents the size of the knapsack}
}
\value{
A list with two elements: the maximum value and a vector of selected items
}
\description{
Solve the knapsack problem using dynamic programming
}
