% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicProgramming.R
\name{knapsack_dynamic_programming}
\alias{knapsack_dynamic_programming}
\title{Solve the knapsack problem using dynamic programming}
\usage{
knapsack_dynamic_programming(weights, values, W)
}
\arguments{
\item{weights}{A vector of item weights}

\item{values}{A vector of item values}

\item{W}{Maximum weight capacity of the knapsack}
}
\value{
A list with two elements: the maximum value and a vector of selected item indices
}
\description{
Solve the knapsack problem using dynamic programming
}
