% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Sample data for Knapsack objects.}
\format{
\subsection{\code{knapsack_objects}}{

A data frame with 2000 rows and 2 columns:
\describe{
\item{w}{weight of each item}
\item{v}{Corresponding value of each item}

...
}
}
}
\usage{
knapsack_objects
}
\description{
Sample data for Knapsack objects.
}
\keyword{datasets}
